#742 Adds factory for EssentialsRoomCombiner
