+++ PepperDashEssentials/PepperDashEssentials.csproj
@@ -149,7 +149,8 @@
     <Compile Include="Room\Types\EssentialsNDisplayRoomBase.cs" />
     <Compile Include="Room\Config\EssentialsRoomConfig.cs" />
     <Compile Include="Room\Types\EssentialsTechRoom.cs" />
-    <Compile Include="Room\Types\IEssentialsHuddleSpaceRoom.cs" />
+    <Compile Include="Room\Types\Interfaces\IEssentialsHuddleSpaceRoom.cs" />
+    <Compile Include="Room\Types\Interfaces\IEssentialsHuddleVtc1Room.cs" />
     <Compile Include="UIDrivers\Environment Drivers\EssentialsEnvironmentDriver.cs" />
     <Compile Include="UIDrivers\Environment Drivers\EssentialsLightingDriver.cs" />
     <Compile Include="UIDrivers\Environment Drivers\EssentialsShadeDriver.cs" />
