--- PepperDashEssentials/Room/Types/IEssentialsHuddleSpaceRoom.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Crestron.SimplSharp;
-
-using PepperDash.Essentials.Core;
-using PepperDash.Essentials.Core.DeviceTypeInterfaces;
-using PepperDash.Essentials.Room.Config;
-using PepperDash.Essentials.Core.Devices;
-using PepperDash.Essentials.Devices.Common.Codec;
-using PepperDash.Essentials.Devices.Common.VideoCodec;
-using PepperDash.Essentials.Devices.Common.AudioCodec;
-
-
-using PepperDash.Core;
-
-namespace PepperDash.Essentials
-{
-    public interface IEssentialsHuddleSpaceRoom : IEssentialsRoom, IHasCurrentSourceInfoChange, IRunRouteAction, IRunDefaultPresentRoute, IHasDefaultDisplay
-    {
-        bool ExcludeFromGlobalFunctions { get; }
-
-        void RunRouteAction(string routeKey);
-
-        EssentialsHuddleRoomPropertiesConfig PropertiesConfig { get; }
-
-        IBasicVolumeControls CurrentVolumeControls { get; }
-
-        event EventHandler<VolumeDeviceChangeEventArgs> CurrentVolumeDeviceChange;
-    }
-
-    public interface IEssentialsHuddleVtc1Room : IEssentialsRoom, IHasCurrentSourceInfoChange,
-        IPrivacy, IHasCurrentVolumeControls, IRunRouteAction, IRunDefaultCallRoute, IHasVideoCodec, IHasAudioCodec, IHasDefaultDisplay, IHasInCallFeedback
-    {
-        EssentialsHuddleVtc1PropertiesConfig PropertiesConfig { get; }
-
-        void RunRouteAction(string routeKey);
-
-        IHasScheduleAwareness ScheduleSource { get; }
-
-        BoolFeedback InCallFeedback { get; }
-
-        BoolFeedback PrivacyModeIsOnFeedback { get; }
-
-        string DefaultCodecRouteString { get; }
-    }
-}
