+++ essentials-framework/Essentials Core/PepperDashEssentialsBase/Room/Combining/EssentialsRoomCombiner.cs
@@ -46,9 +46,12 @@ namespace PepperDash.Essentials.Core
 
             CreateScenarios();
 
-            SetupPartitionStateProviders();
+            AddPostActivationAction(() =>
+            {
+                SetupPartitionStateProviders();
 
-            SetRooms();
+                SetRooms();
+            });
         }
 
         void CreateScenarios()
@@ -242,5 +245,22 @@ namespace PepperDash.Essentials.Core
         }
 
         #endregion
+    }
+
+    public class EssentialsRoomCombinerFactory : EssentialsDeviceFactory<EssentialsRoomCombiner>
+    {
+        public EssentialsRoomCombinerFactory()
+        {
+            TypeNames = new List<string> { "essentialsroomcombiner" };
+        }
+
+        public override EssentialsDevice BuildDevice(PepperDash.Essentials.Core.Config.DeviceConfig dc)
+        {
+            Debug.Console(1, "Factory Attempting to create new EssentialsRoomCombiner Device");
+
+            var props = dc.Properties.ToObject<EssentialsRoomCombinerPropertiesConfig>();
+
+            return new EssentialsRoomCombiner(dc.Key, props);
+        }
     }
 }
